audit-monkee/projects/tools/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
alembic==1.13.2
pydantic==2.9.2
celery==5.4.0
redis==5.0.8
httpx==0.27.2
jinja2==3.1.4
PyNaCl==1.5.0
PyJWT==2.9.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==4.9.4
reportlab==4.2.2
